Relocate 'Criteria' to helpers.py
The change to make METADATA_KEYS a set means we can only have one
definition of the Criteria class. It also eliminates some duplicate
code.
