+++ github/branches/validate_compliance.py
@@ -4,31 +4,10 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from typing import Any, List, Optional, Tuple
-from dataclasses import dataclass
+from typing import List
 
 from .retrieve_github_data import RepoBranchProtections, BranchProtectionRule
-
-
-@dataclass
-class Criteria:
-    standard_number: str  # as defined in messages file. alpha-numeric
-    slug: str  # id to match. alpha-numeric
-    description: str  # Description of invalid
-
-    @classmethod
-    def metadata_to_log(_) -> List[str]:
-        return ["standard_number"]
-
-    @staticmethod
-    def idfn(val: Any) -> Optional[str]:
-        """provide ID for pytest Parametrization."""
-        if isinstance(val, (Criteria,)):
-            return f"{val.standard_number}-{val.slug}"
-        return None
-
-    def __str__(self: Any) -> str:
-        return f"{self.standard_number} {self.description}"
+from ..helpers import Criteria
 
 
 # define the criteria we care about. Identify each critera with a string that will
