+++ github/cli-wrapper.py
@@ -17,18 +17,20 @@ sys.path.append(str(Path(__file__).resolve().parents[1]))
 from github.orgs import retrieve_github_data as org_retrieve_github_data
 from github.branches import retrieve_github_data as branch_retrieve_github_data
 
-sub_command = sys.argv[1]
-del sys.argv[1]
-if sub_command == "orgs":
-    # org will get metadata orgs if none supplied
-    org_retrieve_github_data.main()
-elif sub_command == "branches":
-    # branch does not have default, so pass along current command line
-    # N.B. since that will also happen in pytest's doctest mode, that
-    #      special case is dealt with in the parse_args function
-    branch_retrieve_github_data.main()
-elif sub_command == "--doctest-modules":
-    # Another special case
-    pass
-else:
-    raise SystemError(f"Unknown sub command '{sub_command}'")
+if __name__ == "__main__":
+
+    sub_command = sys.argv[1]
+    del sys.argv[1]
+    if sub_command == "orgs":
+        # org will get metadata orgs if none supplied
+        org_retrieve_github_data.main()
+    elif sub_command == "branches":
+        # branch does not have default, so pass along current command line
+        # N.B. since that will also happen in pytest's doctest mode, that
+        #      special case is dealt with in the parse_args function
+        branch_retrieve_github_data.main()
+    elif sub_command == "--doctest-modules":
+        # Another special case
+        pass
+    else:
+        raise SystemError(f"Unknown sub command '{sub_command}'")
